UART Tx Updated
So, I removed LUT optimization, and replaced interface. 
Here, you can see that one strobe and one flag are added.
flag_txe,- means tx buffer is empty and new data can me uploaded. This means, that 8-bit of data are trnsferred via TX, line. This flag is set on start of stop bit transmission.
flag_txc - this is storbe. This strobe is generated every time when byte is transmitted. Durion of this signal is one i_clk.